vo: create ADQL validator object only when required
Rearrange things so that the TAP query parser only creates an
AdqlValidator object when the list of available tables has
(maybe) changed, rather than doing it every time a validation
operation is performed.

This is a worthwhile optimisation, since creating a validator
may be somewhat expensive for a large TAP service (VizieR's 30k tables),
and validation happens every time a character changes in the ADQL
text field.  Editing queries relating to a large service is
no longer sluggish.