topcat: run in PLASTIC mode only on explicit request
Change topcat startup behaviour so that it only runs in PLASTIC
interop mode if the -plastic flag is specified on the command line.
Otherwise it uses SAMP.

Previously if a ~/.plastic lockfile was detected it would run in
PLASTIC mode.  This happened even if no PLASTIC hub was running,
so a rogue ~/.plastic file would more or less disable SAMP.

One could envisage some smarter logic to check if a PLASTIC hub
is actually running, but I don't think anybody uses PLASTIC any
more at all, so I don't think the additional complication is
justified.