vo: include child tables of included schemas in masked tree
When the Find box is used to mask entries in the table metadata
tree in the TAP query panel, fix it so that when a schema is
included, all its child tables are included as well.
Previously the child tables were only included if their name
matched the find string.  For (most) TAP services this has no effect,
but for those (TAPVizieR) where the table names are not explicitly
qualified by schema, it does make a difference.  Doing it like this
makes the behaviour uniform between the two different cases.