vo: present ADQL text editing in multiple tabs
The ADQL text entry area in the TapQueryPanel is now in a JTabbedPane,
and you can add and remove multiple tabs and switch between them in
order to have several queries on the go at once.
This can be used as a crude history mechanism.

At this commit, there is no GUI to add or remove the tabs.