vo: introduce OutputFormat class to represent TAP output format
TapCapability now has a getOutputFormats() method that returns
an array of OutputFormat objects.  This makes it more convenient
to find out the declared TAPRegExt output formats
(see TAPRegExt v1.0 section 2.4).  It also allows simplification
of the taplint output format checking, since the code in the VO
package does not need to be rewritten there.