vo: allow choice of preferred output FORMAT from TAP load dialog
TapTableLoadDialog now has a setPreferredOutputFormat method to
request a named (VOTable) output format for receipt of TAP result
tables, if it is available.

The Topcat version of the dialogue additionally provides a user
menu to allow choice between some known options (tabledata, binary,
binary2 or default).  The default is currently BINARY2.