vo: launch browser on user click for TAP reference URLs
Add machinery to the MetaPanel to allow a pluggable action to take
place when the user clicks on fields marked as URLs.
The fields are marked in a visible way (blue link, different cursor).

Actual browser launching is pluggable; it's done using the third-party
BrowserLauncher2 in the topcat application, it can be done (slightly
less reliably) using java.awt.Desktop at java 1.6+.