vo: read DALI examples from TAP /examples endpoint
The TAP client now reads examples from the /examples endpoint if
available, and adds relevant entries to a Service-Specific sub-menu
of the Examples popup menu.

The details of the examples format are still somewhat under
discussion (no DALI/TAP examples clients exist up till now AFAIK,
so Markus and I are iterating on the details of how the protocol
does/should work), so implementation may change in the future.