vo: TableMeta/ColumnMeta names now considered ADQL-ready
The TableMeta and ColumnMeta getName() methods are now considered to
return strings suitable for insertion into ADQL queries without
any further quoting or escaping.  Previously, they were considered
unquoted values that should be quoted if they did not follow
the ADQL syntax rules.

This change follows a discussion on the dal@ivoa.net list in
April/May 2015 (with the subject "table_name syntax").
The intention of the TAP and VOSI standards
(previously misinterpreted by me) has apparently always been
that table and column names reported by a TAP service are
in a form ready for use, so they should not be further changed
by clients.

The behaviour of taplint changes somewhat as a result.  Not only
are reported table/column names that are not syntactically suitable
for immediate ADQL use flagged as errors, but some tests involving
presenting queries in which presented column names were further
quoted have been withdrawn.

Since services do not all follow this prescription perfectly,
some things that used to work now don't, as a consequence of
service non-compliance (especially TAPVizieR, whose tables all
have weird names).  This applies mainly to the example ADQL
queries provided in the TapTableLoadDialog.