vo: add TAPVizieR-specific TapMetaReader
At time of writing, VizieR provides a non-standard two-stage
/tables endpoint; the basic document gives you a list of schemas
with tables, but no columns or foreign keys.  Appending the table
name gives you the populated <table> element.

The details of the protocol seem to be in flux at the time of
writing, so some adjustments to the implementation may be required
in the future.