vo: add TableSetSaxHandler methods to extract naked tables
You can now use TableSetSaxHandler to pull out a flat list of all
tables as an alternative to the hierarchical list of schemas
containing tables.  This will pick up tables outside of <schema>
elements too.  It means the same class can be used for TableSet
documents with top-level <table> elements.