ttools: allow dropped quotes in taplint result column names
Some taplint tests check that the column names in a table result
match the column names declared in the table metadata.
As Markus pointed out, these should not require column names
declared as delimited identifiers in table metadata (i.e. with
quotes around the names) to appear with the quotes when they
are used as column names in the result table.  So normalise
column names during this comparison to allow quoted names to
match unquoted ones.