vo: support UDF and geometry func specifics in ADQL validation
The AdqlValidator class can now be constructed in a way that
is specific to a TapLanguage instance, in particular taking
account of the optional user-defined functions and ADQL geometry
functions supported by the language (service) in question.

This works, but there are some issues with its operation related
to the ADQL parser library (1.3).

  - parser errors relating to these declarations
    (UnidentifiedFunctionExceptions) currently lack text position
    information, so it's not easy to report the errors visually

  - UDFs seem to be sensitive to the types of columns, even if
    those columns have been constructed in a "typeless" way
    (DefaultDBColumn constructor "type" argument = null),
    leading to parse errors for legal UDF invocations.

For this reason, don't actually use this facility in the
TapTableLoadDialog just yet.