ttools: update JEL to v2.0.2
I was seeing a problem with JEL (0.9.11) where it was not possible to
compile an expression with a void return value.  This can be required
when supplying JEL activation actions.

This is apparently fixed in JEL 2.0.2, though I don't see anything
in the release notes which explains why.  Upgrade anyway.

Downloaded jel-2_0_2.zip from http://www.gnu.org/software/jel/.
I recompiled from source to get a java5 version with debugging.
The distribution contains jel_g.jar (with debugging) or jel_jdk5.jar
(for java5) but no jar file with both.  Just building from source
with a java5 JDK seems to do the trick (though it would also be
possible to tweak the build file to cross-compile using a later JDK).
I've called the result jel_g.jar, but it's not identical with the
distributed version.  Debug isn't essential, but I'd say it's nice
to have.

Small change to API also required (Hashtable->HashMap in Library
constructor).