topcat: emit activation messages through logging system
Till now, Activation Actions caused a message to get writtten to
System.out using a println.  Changed so that these messages are
now sent through the logging system as INFOs.  They are a little
bit more informative too.  This is more consistent with the rest
of the application and tidier, but it does consitute a visible
change of behaviour, so might possibly cause complaints.

The System.exec() functions however now do write their output
to stdout/stderr direct rather than providing the text as a
return value of the activation function.