vo: improve TAP window ADQL Examples menu
The Examples menu is now multi-level.  The top level contains only
submenus.  As well as its former content, it now has sub-menus
relating to the ObsTAP and RegTAP data models, enabled if those models
are declared by the current TapCapabilities object.

The interface for adding items to this menu is also simplified,
which means the topcat sub-class of TapTableLoadDialog needs to
work a bit less hard.