vo: allow TapSchema metadata reader to pre-load foreign keys
Provide an option in TapSchemaTapMetaReader to load all the
foreign key information in one go rather than doing it as
required once for each table.  This is likely to be much more
efficient if a few tables are being examined, since there are
typically not very many foreign keys in a whole TAP service,
and querying for the keys for each table is relatively expensive.

The AUTO TapMetaPolicy counts the number of foreign key links,
and decides whether it's a good idea to do this.
New policy instances are introduced that preload all columns and
foreign keys, or just all columns, or nothing.