ttools: configure JEL null primitive dereferencing
When JELRowReader has to pass an integer or boolean value to the
JEL expression evaluator, and that value is a null reference
(for a Long/Integer/Short/Byte/Boolean wrapper class object)
there's a problem.  Previously it passed a zero value, evaluated
the function, and then threw away the result.  It did that because
throwing exceptions is expensive.

But sometimes, if there are side-effects of the function evaluation,
that's bad, because the function gets called with the wrong arguments
(zeros).  So add a flag to configure it so it does handle
this situation by throwing a NullPointerException, so the function
doesn't get called at all.

That flag is set for topcat activation actions, where the side-effects
are important, but not for topcat/stilts synthetic columns etc,
which are assumed to be basically functional (side-effect-free).