vo: withdraw assumption about rr.interface intf_index type
When retrieving values from the intf_index column of table
rr.interface in the RegTAP data model, no longer assume that
it has a numeric type.

Although RegTAP v.10 sec 8.8 does tabulate this column as
having data type SMALLINT, that's an error/inconsistency
in the standard, since sec 3 says that the data type is
implementation defined for *_index columns.  Markus Demleitner
confirms that implementation-dependent is the intention
(and anyway relaxing the assumption does no harm).