vo: add TapServiceFinder framework
Add interfaces, GUI and one implementation for locating TAP
services by subject.  Subject currently means keywords matched
against table names or descriptions, though this could be
extended to e.g. column UCDs etc.  This is as distinct from
the search by registry resource record (service name, ivoid etc)
used for the generic DAL table load dialogues.

These classes work, but are not yet integrated into the
TapTableLoadDialog.  The only TapServiceFinder implementation
so far is based on the (non-standard) GLOTS tables hosted at
GAVO DC.  This kind of information should really be sought from
the IVOA Registry, but at time of writing the registry doesn't
have enough information in it (it's not in general fine-grained,
i.e. doesn't contain table metadata for all, or even most,
TAP table datasets), while GLOTS has exactly what we need.
However, the classes are so structured that it should be easy
to slot in a registry-based implementation if/when the registry
becomes capable in practice of supplying the relevant information.