vo: minor adjustments to registry search UI
In the registry query panel, the "Submit Query" button is
renamed to the less confusing "Find Services".

In the TAP load dialogue, the "Enter Query" button is renamed
to "Use Service".

This does have an effect on the appearance of the various
registry-based TableLoadDialog windows; screenshots and possibly
related text documentation should be updated (but have not been
in this commit).