vo: improve table count reporting in TAP metadata tree
Schema nodes in the Tap metadata tree now report the number
of table children in the tree as well as the total number of
tables in the schema (as "<ntPresent>/<ntTotal>"),
if they may be different.  They may be different if a search
term restriction mask is currently applied to the display tree.