table: adjust heuristics to spot header row in CSV files
CSV files may or may not contain a header row, i.e. the first line of
the file may be column names, or may be the first line of the data
proper.  STIL uses a heuristic to try to work this out: if the
first line can be interpreted as a row with data types consistent
with the data types inferred from looking at all the rest of the
rows in the table, it's assumed to be data, otherwise it's assumed
to be column names.

Previously, blank values (zero-length string entries) were treated
differently when assessing the header line than when inferring
column types.  Now, the same criteria are used, so if the first row
contains empty elements, it can still count as a data row.