vo: sort tables and schemas for CADC and /tables meta readers
The Cadc and TableSet implementations of TapMetaReader now
sort the schema and table lists they have read before returning them,
in the same way as the TAP_SCHEMA implementation(s).
This is more consistent.  It's probably the right thing to do,
though in principle services might supply TableSets from the
/tables endpoint in a sequence they would rather have preserved
than re-ordered.