vo: use JComboBox for service URL selection in TAP load dialog
Use an ExampleSelectField for choosing the TAP service URL in
the TAP table load dialog.  This provides example text before use
as before, but it also stores previously selected values making it
easier to flip between TAP services.

This does mean that the Use Service tab can't be disabled,
but I've added a popup if an attempt to do that is made when
there is no selected URL.