vo: fix (newly?) broken VizierTapMetaReader
I'm pretty sure it used to work, but when I tried it today it didn't,
so I think the service has changed.  I think the service used to
provide table names without schema prefixes, and now it includes them,
so the table names I was using to request table metadata were not
recognised.