topcat: prevent submission of null-SIZE multi-cone/SIA queries
If you erased the default value of the Search Radius field in
the multi-cone/multi-SIA window, it would send illegal multiple
queries to servers.  There's no sense in this.  Fix it so that
if that field is blank, the GUI gives you an error popup rather
than sending the queries.  Implementation slightly complicated
(which is presumably why I did it this way in the first place)
by the fact that it *does* make sense to specify a blank SIZE
for multi-SSA queries.