vo: switch on UDF-sensitive ADQL parsing
Where User-Defined Functions are declared by the TAP service,
use the declared list to configure the ADQL validator.
Misuse of UDFs or use of unknown UDFs can thus be flagged
as parse errors.

The code changes to the starjava code are minimal, but a modified
version of the ADQL parser with two fixes (error position information
for UnresolvedFunctionExceptions and relaxation of datatype matching
rules) is required to make it work properly.  These fixes were
kindly implemented by Gregory Mantelet at my request.

The adqlParser library was built using my version 01135a4c491a,
which is a minor tweak (for building with starjava) of Gregory's
b49f81608d3.