ttools: fix votcopy handling of unsignedByte columns
Votcopy was changing the datatype for unsignedByte columns to type
short during copying, on the assumption that this is what the
VOSerializers do (because unsigned bytes are hard to handle in java).
They used to do that, but they don't any longer, since the
introduction of Tables.UBYTE_FLAG_INFO (cc7abd5d59)
previous few commits).

The result was slightly annoying for TABLEDATA output, since
the datatype was unnecessarily changed.  For BINARY/BINARY2
output, it generated unreadable VOTables.

Remove that bit of code, so votcopy works for these columns again.

The code is badly designed, or this sort of thing couldn't happen.
The VOTable serialization code really needs rewriting so that
output column types can be determined programmatically rather
than requiring expert knowledge at different places in the code.

But at least there is test code there now that should catch this
sort of thing for votcopy.