votable: fix datatype="bit" binary stream read bug
The code for reading bit-vector columns from a BINARY/BINARY2
stream was completely wrong.  It suffered from the same error
as present in the recently-fixed FITS TFORMn='rX' bug, but
additionally read the wrong number of bytes from the stream
(equal to the number of bits it wanted).  This was more drastic,
but would almost always lead to a total read failure of the
file, so the error would be obvious, unlike the FITS issue.

Given that this bug has been present for ever, and nobody has
reported it before now, I can only assume that nobody is using
STIL for binary-serialialised VOTables with datatype="bit"
columns.