vo: improve TAP query error reporting
TAP says that error messages have to be returned as VOTable documents
(which is really a pain).  Inevitably, this is not always the case,
since the error message may be generated at the HTTP rather than
TAP service level, so we have to cope with non-VOTable error
responses too.

Previously, non-VOTable error responses just got thrown as
exceptions with the message "TAP response is not a VOTable",
which makes it difficult for developer or user to know what's wrong.
In this commit, try to do better.  For non-VOTable error responses,
try to grab the HTTP response code, message and (at least part of)
body and present those as the message of the IOException that gets
thrown, presumably for eventual presentation to the user.