vo: log curl(1) equivalent of TAP queries at CONFIG level
When a UWS job query is executed with HTTP POST (UwsJob.postForm)
an approximately equivalent curl(1) command is now generated
and logged at the CONFIG level.  This can be very useful when
trying to investigate recalcitrant TAP services and pass the
findings back to service providers.  The curl output isn't
guaranteed bulletproof or identical to the java implementation,
but it's a good start.