vo: implement RegTAP+aux -based TapServiceFinder
Add a new TapServiceFinder implementation that uses the
relational registry along with the #aux-fragment conventions
suggested in an early version of Markus Demleitner's IVOA Note
"Discovering Data Collections Within Services".

A certain amount of trial and error was used to get it working,
but it seems OK.  However, there are very few compliant records
in the registry at the time of writing, so it's hard to say how well.

It is possible that the ADQL this class uses for its queries
admits of some improvement.