vo: cope with DALI- *or* TAPNote-style TAP service examples
The details of how to mark up example ADQL queries in RDFa
at the /examples endpoint of a TAP service are currently
in dispute.

DALI v1.0 says do something like this:

   <div property="generic-parameter" typeof="keyval">
      <span class="invisible" property="key">QUERY</span>
      <pre property="value" >SELECT * FROM table</pre>
   </div>

while the TAP Implementation Note v1.0 says this:

   <pre property="query">SELECT * FROM table</pre>

In absence of a ruling, take a belt-and-braces approach.
Look for either form, and use whatever's there.
The TAPNote-style one is preferred if both happen to be present.