vo: fix ADQL parse bug related to delimited column names
If a column was referenced in ADQL text as either 't."c"' or
't.c' where '"c"' is a delimited identifier declared as the column
name and t is a declared table name, the parser was throwing a
NullPointerException.  This was because the DBTable implementation
was set up wrong in terms of how it was reporting its column names
(though the parser's behaviour in throwing an unchecked exception
here is also suboptimal).

Fixed the DBTable implementation in AdqlValidator following advice
from Gregory Mantelet.