ttools: display non-absolute colour ramps differently
Colour ramps are now displayed for non-absolute shaders with
the effect of the shader on a vector of colours (itself supplied
by an (absolute) shader), rather than a fixed colour.
This can look slightly messy, but non-absolute shaders are messy;
it is probably a good idea to signal this in the GUI for least
surprise (rather than just marking them with a "*" as at present.

This allowed to replace the createIcon instance method in the Shader
interface with a static utility method in the Shaders class.
This is much tidier and requires less casting to known types and
repeated implementation.