ttools: fix transparency shader
The "Transparency" Shader that altered alpha was not working
correctly for Aux shading, and I don't think ever had done.
It's because the LayerOpt was set to opaque if the Aux transparency
was turned off, on the assumption that the shader colours were all
opaque.  Fix it so that the LayerOpt is sensitive to the opacity
of the shader as well as the full-range transparency.

Do the same for the Density shading mode.  This has not up till now
had a transparency shader, but I think it could do.