votable: tweak VOTable DOM getParentNode implementation
The getParentNode implementation used when constructing the
VOTable DOM was returning a top-level element that was not
identical with the return value of getOwnerDocument.
This caused trouble with XPath implementations
(java.lang.RuntimeException: Could not resolve the node to a handle
        at
com.sun.org.apache.xml.internal.dtm.ref.DTMManagerDefault.getDTMHandleFromNode(DTMManagerDefault.java:556)

Fix the getParentNode implementation so that (as a special case) it
returns the owner Document when reporting the top-level parent.
This is a bit hacky, but should probably be OK (delegation of all
DOM node implementation is fiddlier than it sounds).

Thanks to Omar Laurino for doing much of the leg-work to track this
down.