ttools: implement sky density plotter
Add a new plotter that plots a density map on a SkySurface.
It currently uses a HEALPix grid.

This is a true density map, unlike the existing arrangements that
let you plot markers with density-sensitive shading.
The density values may be weighted by an additional coordinate.
Other features may be added in later revisions.