cdf: fix test to work in non-UTC timezones
This bug interacts with a butg relating to incorrect time zone
handling in JCDF.  When the JCDF version in starjava is updated
to a version that fixes that bug, make sure this starjava fix
is applied at the same time, or the test will fail.
It can't be applied before then.