ttools: limit SkyDensityPlotter pixel resolution
Don't sample the density at a finer resolution than it's going to be
displayed.  That is not only unnecessarily expensive, it also
results in a worse plot, since the resampling is done by pulling
values from the sky bin grid to the scren pixel grid, so many
sky bin grid values get missed and the plot misses much of the data
(so typically looks noisy).

I experimented with adding an alternative resampling mode which
drizzles values from the sky bin grid onto the screen pixel grid,
but this is likely to be expensive and it adds a reasonable amount
of complication to the IntegerBinBag API and implementation;
I think this is a better solution.