topcat: add vector TOPCAT logo
This logo is the one drawn by Phil Hall in 2012 for TOPCAT.
It's more or less the PDF original, but I clipped the borders off
using some interactive tool I forgot the name of.
This resulting file is much smaller than the supplied original
(10kB vs. 110kB).  I can't see the difference (it certainly doesn't
seem to have been rasterised), so hopefully what's lost is just
some kind of invisible metadata or history information.

This PDF logo is not currently referenced by the documentation
or build, but it's added to the repository at the (eminently sensible)
suggestion of Paul Sladen.