ttools: add example/BasicPlotGui class
Provide source code for a minimal plot application that uses the
self-describing plot2 API to present the user with GUI controls
to configure a plot.

As it stands this is unsophisticated in various ways: only a single
fixed plot layer is offered, and replots (though not navigation
operations) require an explicit replot action by the user.
It may get souped up in some future version.