ttools: add Object$colid syntax to JEL expression language
Add a new syntax element Object$<column-identifier> to the JEL
expression language.  It yields the column value as an Object
and not a primitive.

This can be useful if you're passing the values to a function
which needs to know whether it has nulls or not.  The normal
column decoding behaviour maps column values to primitives,
and just bails out if an attempt is made to dereference a
null-valued integer or boolean object.

This added at the request of Celine Charavay (CEA)