ttools: modify AuxReader interface
The range adjusting (sole) method of the AuxReader interface is
now changed so that when it's called it is expected to iterate
over an entire TupleSequence rather than getting called once
for each tseq entry.  This allows more flexibility in the
implementation: you can update the range based on global properties
of the data rather than something that only is affected by each
individual entry.

The method arguments are the same, but the method name has been
changed from updateAuxRange to adjustAuxRange, just to make sure
that code understands there's a difference.