ttools: replace IntegerBinBag with BinList
The BinList class does more or less what the (poorly-named)
IntegerBinBag class did, but there are some changes to the
interface and implementation.

A Combiner instance is now responsible for creating its own
array-based BinList implementation if it wants to,
rather than providing an interface that allows generic array-based
BinLists to be built from it - this results in code that is less
complex and more efficient (especially for MEAN).

At the same time add new COUNT and HIT instances to Combiner.