ttools: SkyDensityPlotter now uses global aux scale
The SkyDensityPlotter now uses AuxScale.COLOR, the global Aux
colour ramp, rather than allowing users to specify per-layer
values for colour map, stretch, limits etc.
This gives more control and more quantitative information about
what the plot is doing.  But it comes at a cost, because the
aux scale ranges (typically) have to be worked out before the plot
is made, duplicating a lot of the work that the plot does.
Pros and cons.  It wouldn't be too hard to add a config option
(at least in the Plotter instance constructor) to switch this.