ttools: improve pos/neg switch in SpinnerPairSpecifier
Fix the IntegerConfigKey.SpinnerPairSpecifier so that the two
(positive and negative) spinners can be kept in sync with each other.

This is currently (only) used in the HEALPix density plot:
it means that if you change the HEALPix level by changing whichever
control is active, the other (inactive) one gets updated to a value
that means the same thing as you do it.  That means that flicking
between Absolute and Relative mode for this level specifier doesn't
(usually) affect the actually displayed plot.  That makes a smoother
user experience.