ttools: add unit test that reveals a crossmatch bug
A bug in crossmatching has been reported by Grant Kennedy (IoA).
It applies to MatchEngines (match criteria) that have per-object
errors, and it relates to the bounds restriction that is done
at the start of a match to cut down the number of object that
are binned to those that might match each other.

It seems that matches can be missed near the edge of the
restricted intersection region in some cases, I'm not yet
completely sure why.

The newly-added unit test fails at this commit.  The test is
added to ttools, but the bug is in the table.join package.