ttools: fix title painting bug for exported plot bitmaps
When exporting (from topcat) or generating an output file
(from stilts) of a plot, the colour of the title was not
being set explicitly.  This meant that although it was visible
in PDF exports, for bitmapped formats (PNG, GIF, JPEG) the
characters were getting painted in white on white, which
normally means they are invisible.