table: change signature of MatchEngine.getMatchBounds method
Replace Comparable[][] with Range for the parameters and return
value of MatchEngine.getMatchBounds.  This doesn't change any
behaviour, but it's going to make it easier/less confusing
to do what I need to do next, which is to pass not just one
range but multiple ranges into this method for evaluation.

This is a backwardly incompatible change to the public API of STIL
(expect others to the same method in subsequent commits).
However, I don't *think* anybody else uses this method in
application code, so it's probably not too harmful.
The classes in the table.join package are really only semi-public
anyway, since they are not properly documented except in the javadoc.