table: fix crossmatch range-restriction bug
There was a significant bug in the crossmatching, relating
to the optimisation step that restricts the range over
which crossmatches will be sought.

Prior to the actual crossmatching, the RowMatcher attempts to locate
the regions in which actual associations might exist, so that it
can limit the amount of unnecessary work it does to regions in
which all the relevant tables actually have coverage.
It does this (where possible) by asking the MatchEngine to
identify hyper-cubic regions in the N-dimensional tuple space
which bound all the rows in a particular table to be matched.
Then only the intersection of all these bounded regions needs to
be searched.  Of course these hyper-cubic regions have to be
broadened by a little bit to account for the match error allowed,
since a match may occur a match-radius-length outside each of
the bounding regions.  That broadening is done (where possible)
by the MatchEngine.getMatchBounds method (if implemented).

In previous versions, this broadening was done by looking at all
the tuples for the table being assessed and using the maximum
error.  That's OK if the error is fixed, but if the error
depends on table row it's not, since you need to broaden each
hypercube by the largest error of *any* of the matched tables,
not just of the table bounded by that hypercube, otherwise some
matches may get missed at the edges.  This means that some
associations could have been missed when using certain MatchEngines,
especially if the error scales were significantly different
between the matched tables.  The bug would not have generated
any false associations.  Only the following MatchEngines were
affected:

   Class                        STILTS name   TOPCAT name
   -----                        -----------   -----------
   ErrorCartesianMatchEngine    <n>d_err      <n>-d Cartesian with Errors
   ErrorSkyMatchEngine          skyerr        Sky with Errors
   EllipseCartesianMatchEngine  2d_ellipse    2-d Cartesian Ellipses
   EllipseSkyMatchEngine        skyellipse    Sky Ellipses

This commit plugs that hole, by assessing the broadening in the
context of *all* the tables being matched.

Thanks to Grant Kennedy (IoA) for reporting this bug.