ttools: add median combiner for weighted density plot
Per-pixel median calculation is expensive, since it needs to keep a
copy of every submitted data point, but Alex Bombrun says it's an
important requirement for Gaia.

Added a new class QuantileCombiner which could easily be used to
to calculate other quantiles or more general quantities, but for
now the only option presented in the UI is Median.