ttools: add getCoordIndex method to AuxReader
AuxReader objects now (where possible) report the coordinate they
are ranging for.  This provides the opportunity for parts of the
plotting infrastructure to find out from a PlotLayer which auxiliary
coordinates it is performing ranges for.  That in turn allows them
to make a good guess about how to label auxiliary ranges.