ttools: provide default label for Aux axis colour ramp
AbstractPlot2Task now investigates the layers to be plotted,
in order to find out which correspond to the AuxScale.COLOR scale.
If any do, it takes the first one and uses the user-entered
JEL expression from that as a default for the auxlabel parameter.

The effect is that the colour ramp is usually labelled with a
physically meaningful string, rather than being left blank.