topcat: generalise procedure for labelling Aux axis colour ramp
Modify the way that topcat identifies the label for the ShadeAxis.
Previously, it interrogated GUI components directly, based on their
text label.  This was rather hacky, and meant that it wasn't easy
to identify a label based on a "Weight" coordinate as well as an
"Aux" one (which the new weighted modes want to do).

Instead, it now looks at the layers to be plotted that are apparently
using the AuxScale.COLOR scale, and pulls out the text associated
with the relevant coordinate in that layer's DataSpec.  This should
continue to do the same thing (get the same labels) for the existing
Aux layers, but it also gets labels for any other layers that use
(or at least do ranging based on) AuxScale.COLOR.  So it works for
weighted density layers too.