ttools: report information about pixel size from some plots
Weighted density plots (ShapeMode.WeightedDensityMode and
SkyDensityPlotter) now provide some quantitative information
on pixel size in their returned ReportMaps.

I'm not sure this is currently done in its optimal form, and it's
not currently passed to the user in a very helpful way, but this
information is required if you want to do quantitative analysis
for some of the colour-coded output values.