topcat: rationalise plot report logging
All "general-interest" plot reports are now logged at the INFO level
through the logging system from one place (StackPlotWindow), in
addition to other places that information may be presented in the GUI.
The ReportLogger class is no longer used for custom per-control logging.

This is not the final solution for reporting, since logged reports
will only be seen by expert users, but at least this information can
be seen.  Ideally the information would be additinoally presented
in the GUI (as it is in some, but not all, cases already).