ttools: use DensityPlotter in PlanePlotType
There are a couple of things to note.

First, adding the DensityPlotter really means we have too many
icons in the topcat Form selector toolbar - they overflow the default
window width, so you don't see this one unless you resize that
window interactively.  Could make the window wider by default,
but really should rethink this GUI, since more plotter types may
be added in the future.

Second, this plotter could also be used in other plot types (I think):
Sky and Cube.  Sky already has a SkyDensity plotter, which is more
respectable, but though you could imagine wanting this pixel-based
functionality for some reason.  Cube: it just seems a bit weird to
average over pixels that are not aligned with the axes.
For now we stick to plane plot.