vo: rework UwsJob job status interface
Make some modifications to the parts of the UwsJob interface that
have to do with determining the current/recent state of the job.
As it stood, this had a lot of hidden state, and was fragile.
It was also hard to make it work well with the UWS 1.1-style
blocking calls (which apply to reading the job document, but not
the status string on its own).

Although the API is changed here, it should not make any difference
to application behaviour.  However, the code was messy in parts,
so mistakes are possible.