jniast: remove sameObject test after addFrame
The behavior of astAddFrame changed in Starlink commit bc2bc84aab
(Apr 28 2015) so that it now stores copies of the relevant
objects.  So if the current version of AST is installed,
the test that the frame is the same object after adding it to
the frameset fails.  This commit replaces the "sameObject"
comparison with "equals", since the two frames are no longer
expected to be the same object.

(cherry picked from commit 8e5b6331f604610feb04a2a52410de6917df52d1)