ttools: implement Fill plotter
This notionally paints the area under the curve represented by a
set of points representing a function y(x).  If you give it a
set of points that don't represent a (single-valued) function,
it might look strange.

One thing about the current implementation is a bit scrappy -
at high magnification (more pixels across than data points),
the plotted bars are aligned to the right of the corresponding
data point rather than centered on it.  That should probably
be fixed.