ttools: update AbstractPlot2Task to work with a Gang of surfaces
The plot task framework is now capable of working with multi-zone
plots.  Actual support is dependent on the PlotType-specific
AbstractPlot2Task concrete subclass (no multi-zone implementations
included at this commit).  The way it works is that a new zoneN
parameter may be specified for each layer N, and layers sharing
the same zoneN value are grouped together into a single zone.

Some parameters have now acquired a zone-specific suffix Z,
so that they can be set per-zone or globally in the same way
as with layer suffixes.  This involves rewriting some of the
code to generalise the idea of parameter suffixes a bit,
since we now have both layer suffixes and zone suffixes.
There may still be a few parameters that could or should be
zone-specific (fonts?) but are not.

Although there are quite a lot of changes to the AbstractPlot2Task
implementation (essentially, adding a new loop over zones where
appropriate), the user interface (command line parameters) should
remain unchanged for the single-zone case, either where zones
are not supported, or where no zoneN parameters are supplied
meaning that all layers live together in the same unnamed zone.