topcat: introduce ZoneDefiner interface
Group the various zone-dependent factories that are used for PlotPanel
construction (by StackPlotWindow) into a single new interface,
ZoneDefiner.  This is just a refactoring operation which does not
change behaviour.  It tidies up the API a little bit, but the main
purpose is to prepare the way for multi-zone (ganged) plots in PlotPanel,
which will use multiple ZoneDefiners rather than just one to specify
gang configuration.