topcat: rework PlotPanel for multi-zone plots
Rework PlotPanel interface and implementation so that it can handle
a gang of related plot surfaces rather than just one.
This is quite fiddly, since PlotPanel is one of the most complicated
parts of the plotting machinery.  Errors here may not show up except
as slower plotting in certain circumstances.  I've done a certain
amount of testing, but look out for regression performance issues.

At this commit, the harnes classes StackPlotWindow and
HistogramPlotWindow are still hardwired to use a single zone.